Type Annotation in disk_gradient

+++ src/disk_gradient.jl
@@ -1,6 +1,6 @@
 
 
-function disk_ab_v(u::Array{Float64,2}, k)
+function disk_ab_v(u::Array{Float64,2}, k::Int)
 	n = size(u,1)
 	m = size(u,2)
 	h = 1/(n*m)
